+++ context/log/index.js
@@ -1,9 +1,50 @@
 'use strict';
 
-const initializeLogger = () => {
+module.exports = (c) => {
+  const serializers = pino.stdSerializers;
+  const redact = {
+    paths: [
+      c.s.redactedLogPaths,
+      'event.Records[*].dynamodb',
+    ],
+    remove: true,
+  };
 
-};
+  const log = pino({
+    level: c.s.config.level,
+    redact,
+    serializers,
+    base: null
+  });
 
-module.exports = {
-  initializeLogger,
-};
+  return async (event, c, next) => {
+    const {
+      awsRequestId,
+      functionName,
+      functionVersion,
+      requestId,
+    } = c;
+
+    const xRayTraceId = c.d.xray.getXRayTraceId();
+
+    Object.assign(c, {
+      log: log.child({ requestId }),
+    });
+
+    c.log.info({
+      functionName,
+      functionVersion,
+      awsRequestId,
+      xRayTraceId
+    });
+    c.log.trace({ event }, 'incoming event');
+
+    try {
+      return await next();
+    } catch (err) {
+      c.log.error({ err }, 'unhandled err');
+      throw err;
+    }
+  };
+}
+
