+++ lamda/service/apiGateway/searchPosts.js
