+++ lamda/service/dynamoStream/streamListener.js
